Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of a char: %zu byte(s)\n", sizeof(a));
-	printf("Size of an int: %zu byte(s)\n", sizeof(b));
-	printf("size of a long int: %zu byte(s)\n", sizeof(c));
-	printf("Size of a long long int: %zu byte(s)\n", sizeof(d));
-	printf("Size of a float: %zu byte(s)\n", sizeof(e));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
 	return (0);
 
 }
